+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## (In development) 5.3.0 (XXXX.XX.XX)
 
 - New: Make it possible to hide latlng field for GeoBlock (@vladox)
+- Fix: Solve issue with address not working streamfield in Wagtail 1.13+ (@vladox)
 
 
 ## 5.2.0 (2022.01.04)